adds static copy, merges build dev/prod, sourcemap
