+++ template/.prettierrc
@@ -1,8 +1,4 @@
 {
-    "printWidth": 110,
-    "tabWidth": 4,
-    "semi": true,
-    "singleQuote": true,
-    "arrowParens": "avoid",
-    "trailingComma": "es5"
+    "printWidth": 120,
+    "singleQuote": true
 }
