+++ template/src/main.css
@@ -1,1 +1,4 @@
 /* Add your CSS styles here */
+.myClass {
+  background-color: #fff;
+}
